pyproject.toml
[project]
name = "py-drivel"
version = "0.1.0"
description = "Get fun metasyntactic variable names"
authors = [
    {name = "Tucker Beck", email ="tucker.beck@gmail.com"},
]
readme = "README.md"
license-files = ["LICENSE.md"]
requires-python = ">=3.10"
dependencies = [
    "auto-name-enum>=2.0.0",
    "inflection>=0.5.1",
    "loguru>=0.7.3",
    "py-buzz~=6.0.0",
    "pydantic>=2.10.6",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.2",
    "snick>=1.4.1",
    "toml>=0.10.2",
    "typer>=0.15.2",
]

[project.urls]
homepage = "https://github.com/dusktreader/drivel"
source = "https://github.com/dusktreader/drivel"
changelog = "https://github.com/dusktreader/drivel/blob/main/CHANGELOG.md"

[tool.uv]
package = true


[dependency-groups]
dev = [
    "ipython>=8.34.0",
    "mypy>=1.15.0",
    "pyclean>=3.1.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-random-order>=1.1.1",
    "ruff>=0.11.2",
]

[project.scripts]
drivel = "drivel.cli.main:cli"

[tool.ruff]
line-length = 120

[[tool.mypy.overrides]]
module = ["strictyaml"]
ignore_missing_imports = true

[tool.hatch.build]
include = [
    "src/drivel/**/*.py",
    "src/drivel/assets/**/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/drivel"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
